Move codegenWrapperKernel to a base GpuReductionTester class
